refactor(messenger-chat): extract conversation header
